# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: MSI Hackathon API (YAML)
  description: Дополнительная документация из YAML файла
  version: 1.0.0-yaml
paths:
...
  /api/projects/photos/{projectPhotoId}:
    delete:
      summary: Удалить фотографию проекта
      description: "Удаляет фотографию проекта по её идентификатору."
      parameters:
        - name: projectPhotoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Фотография удалена
        '404':
          description: Фотография не найдена
...
  /api/photos/{id}:
    get:
      summary: Получить фотографию по ID
      description: Возвращает информацию о фотографии по её идентификатору.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Фотография найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhotoResponseDto'
        '404':
          description: Фотография не найдена

    delete:
      summary: Удалить фотографию
      description: Удаляет фотографию по идентификатору.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Фотография удалена
        '404':
          description: Фотография не найдена
...
components:
  schemas:
...
    PhotoResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          description: URL сохранённого изображения
        createdAt:
          type: string
          format: date-time
          description: Время загрузки фотографии
...